Fixing GrainGenerator::m_OverlapAllowed to be an integer in both the GUI and the support code

+++ Source/AIM/GrainGenerator/GrainGenerator.h
@@ -67,7 +67,7 @@ Q_OBJECT
     MXA_INSTANCE_PROPERTY_m(double, XResolution)
     MXA_INSTANCE_PROPERTY_m(double, YResolution)
     MXA_INSTANCE_PROPERTY_m(double, ZResolution)
-    MXA_INSTANCE_PROPERTY_m(double, OverlapAllowed)
+    MXA_INSTANCE_PROPERTY_m(int, OverlapAllowed)
     MXA_INSTANCE_PROPERTY_m(int, OverlapAssignment)
     MXA_INSTANCE_PROPERTY_m(int, CrystalStructure)
     MXA_INSTANCE_PROPERTY_m(int, ErrorCondition);
